+++ src/astronomo/widgets/certificates/__init__.py
@@ -3,11 +3,13 @@
 from astronomo.widgets.certificates.confirm_delete_modal import ConfirmDeleteModal
 from astronomo.widgets.certificates.create_modal import CreateIdentityModal
 from astronomo.widgets.certificates.edit_modal import EditIdentityModal
+from astronomo.widgets.certificates.import_lagrange_modal import ImportLagrangeModal
 from astronomo.widgets.certificates.manage_urls_modal import ManageUrlsModal
 
 __all__ = [
+    "ConfirmDeleteModal",
     "CreateIdentityModal",
     "EditIdentityModal",
-    "ConfirmDeleteModal",
+    "ImportLagrangeModal",
     "ManageUrlsModal",
 ]